src/requirements.txt
arrow==0.12.1
bitstring==3.1.5
certifi==2018.4.16
chardet==3.0.4
future==0.16.0
idna==2.7
iso8601==0.1.12
Pillow==5.1.0
psutil==5.4.6
pypiwin32==223
PyQt5==5.10.1
pyserial==3.4
python-dateutil==2.7.3
pywin32==223
PyYAML==3.12
requests==2.19.1
sip==4.19.8
six==1.11.0
untangle==1.1.1
urllib3==1.23
